+++ kernel-hal/src/bare/arch/x86_64/interrupt.rs
@@ -55,7 +55,7 @@ hal_fn_impl! {
         }
 
         fn handle_irq(vector: usize) {
-            all_irq().first_unwrap().handle_irq(vector as usize);
+            all_irq().first_unwrap().handle_irq(vector);
         }
 
         fn msi_alloc_block(requested_irqs: usize) -> HalResult<Range<usize>> {
