+++ linux-syscall/src/ipc.rs
@@ -170,7 +170,7 @@ impl Syscall<'_> {
                 Ok(0)
             }
             _ => {
-                let sem = &sem_array[num as usize];
+                let sem = &sem_array[num];
                 match cmd {
                     SemctlCmds::GETPID => Ok(sem.get_pid()),
                     SemctlCmds::GETVAL => Ok(sem.get() as usize),
